feat: rename files to `index.ts`
